Rename DashboardUniqueVisitorsWidget to DashboardSearchVisitorsWidget.
