+++ assets/js/modules/analytics/components/dashboard/DashboardSearchVisitorsWidget.js
