+++ assets/js/modules/analytics/index.js
@@ -40,7 +40,7 @@ import { SettingsEdit, SettingsView } from './components/settings';
 import DashboardAllTrafficWidget from './components/dashboard/DashboardAllTrafficWidget';
 import DashboardPopularPagesWidget from './components/dashboard/DashboardPopularPagesWidget';
 import DashboardGoalsWidget from './components/dashboard/DashboardGoalsWidget';
-import DashboardUniqueVisitorsWidget from './components/dashboard/DashboardUniqueVisitorsWidget';
+import DashboardSearchVisitorsWidget from './components/dashboard/DashboardSearchVisitorsWidget';
 import DashboardBounceRateWidget from './components/dashboard/DashboardBounceRateWidget';
 import { ModuleOverviewWidget, ModulePopularPagesWidget, ModuleAcquisitionChannelsWidget } from './components/module';
 
@@ -83,7 +83,7 @@ export const registerWidgets = ( widgets ) => {
 	widgets.registerWidget(
 		'analyticsUniqueVisitors',
 		{
-			Component: DashboardUniqueVisitorsWidget,
+			Component: DashboardSearchVisitorsWidget,
 			width: widgets.WIDGET_WIDTHS.QUARTER,
 			priority: 3,
 			wrapWidget: true,
